fix(writer): move version-retry helper out of writer pkg to keep gateway light
repositories/objects.py is imported by the gateway (via repositories/__init__).
Importing hippius_s3.writer.db from it triggered writer/__init__ -> object_writer
-> cache/crypto/KMS, loading the whole write pipeline into the gateway at startup
and making its container unhealthy (E2E gateway healthcheck failed). Move
retry_on_object_version_conflict to a lightweight top-level hippius_s3/db_retry.py
(asyncpg only) and import it from all four callers.
